requirements.txt
streamlit
sentence-transformers
pypdf
langchain
faiss-cpu
groq
python-dotenv
IProgress